moved

+++ src/main/java/zombienado_v1/server/serverWeapon/ServerAK47.java
@@ -12,7 +12,7 @@ public class ServerAK47 extends ServerWeapon {
     private final static int BULLETSPEED = 20;
     private final static int MAXAMMO = 360;
     private final static int SHOTS = 1;
-    private final static double SPRAY = 0.3;
+    private final static float SPRAY = 0.3f;
     private final static double RATEOFFIRE = 100;
 
     public ServerAK47() {
