+++ src/main/java/zombienado_v1/server/serverWeapon/ServerUzi.java
@@ -12,7 +12,7 @@ public class ServerUzi extends ServerWeapon {
     private final static int BULLETSPEED = 20;
     private final static int MAXAMMO = 500;
     private final static int SHOTS = 1;
-    private final static double SPRAY = 0.5;
+    private final static float SPRAY = 0.5f;
     private final static double RATEOFFIRE = 80;
 
     public ServerUzi() {
